refactor: Extract hierarchy helper
